Post-model check on rtp structure
missing model (ERA/ECMWF/MERRA) files can leave the rtp structure in
an unbalanced state with model-related entries having fewer obs than
non-model entries. The fill_* routines don't have a problem with this
but downstream processing does (and should). Added a quick check after
adding model data to see that nobs is consistent throughout the rtp
struct. It's done here because this was quick to implement and test
but, needs to propagate out to all other rtp routines or be
incorporated in the fill_* routines.