AIRS: updated read_airibrad.m to use high level matlab HDF-EOS functions
old style hdfsw is no longer in matlab. THey have a new OO module version but
we have opted to go with the high level hdfread utilities. There are differences
in how the two interfaces read data arrays (usually just a transpose) but also some
data is read in as matlab cells through hdfread. read_airibrad is reading all previous
data in a form that looks proper but, it is possible there are missed transposes or other
secondary issues that will need to be found and cleared up.

Data that reads as cell arrays is not checked for size of arrays and code assumes the
proper and desired data is in the var{1} cell bin.